Fix pytest init warnings. (#521)
