+++ setup.cfg
@@ -1,2 +1,6 @@
 [metadata]
 description-file = README.md
+[tool:pytest]
+python_classes = *Test
+python_files = *_test.py
+testpaths = tests/mobly
